Crate sam4l

Peripheral implementations for the SAM4L MCU.

http://www.atmel.com/microsite/sam4l/default.aspx

Reexports

extern crate core as core;

extern crate cortexm4;

extern crate kernel;

use core::prelude::v1::*;

Modules

adc
Implementation of the SAM4L ADCIFE.

aes
Implementation of the AESA peripheral on the SAM4L.

ast
Implementation of a single hardware timer.

bpm
Implementation of the BPM peripheral.

bscif
Implementation of the Backup System Control Interface (BSCIF) peripheral.

chip
Interrupt mapping and DMA channel setup.

crccu
Implementation of the SAM4L CRCCU.

dac
Implementation of the SAM4L DACC.

dma
Implementation of the PDCA DMA peripheral.

flashcalw
Implementation of the SAM4L flash controller.

gpio
Implementation of the GPIO controller.

helpers

i2c
Implementation of the SAM4L TWIMS peripheral.

nvic
Implementation of the SAM4L interrupt controller.

pm
Implementation of the power manager (PM) peripheral.

scif
Implementation of the system control interface for the SAM4L.

spi
Implementation of DMA-based SPI master and slave communication for the SAM4L.

trng
Implementation of the SAM4L TRNG.

usart
Implementation of the SAM4L USART peripheral.

wdt
Implementation of the SAM4L hardware watchdog timer.

Macros

interrupt_handler
Define a function for an interrupt and enqueue the interrupt on the global queue.

Statics

BASE_VECTORS

INTERRUPT_TABLE

IRQS

_erelocate

_etext

_ezero

_srelocate

_szero

Functions

SVC_Handler⚠

_estack⚠

generic_isr⚠

hard_fault_handler⚠

init⚠

reset_handler⚠

systick_handler⚠

unhandled_interrupt⚠