Struct sam4l::crccu::Crccu

pub struct Crccu<'a> {
    client: Option<&'a Client>,
    state: Cell<State>,
    alg: Cell<CrcAlg>,
    descriptor_space: [u8; 532],
}

State for managing the CRCCU

Fields
client: Option<&'a Client> state: Cell<State> alg: Cell<CrcAlg> descriptor_space: [u8; 532]
Methods

impl<'a> Crccu<'a>

const fn new() -> Self

fn init(&self)

fn enable(&self)

Enable the CRCCU's clocks and interrupt

fn disable(&self)

Disable the CRCCU's clocks and interrupt

fn set_client(&mut self, client: &'a Client)

Set a client to receive results from the CRCCU

fn get_client(&self) -> Option<&'a Client>

Get the client currently receiving results from the CRCCU

fn set_descriptor(&self, addr: u32, ctrl: TCR, crc: u32)

fn get_tcr(&self) -> TCR

fn descriptor(&self) -> *mut Descriptor

fn handle_interrupt(&mut self)

Handle an interrupt from the CRCCU