Enum sam4l::flashcalw::FlashCMD

pub enum FlashCMD {
    NOP,
    WP,
    EP,
    CPB,
    LP,
    UP,
    EA,
    WGPB,
    EGPB,
    SSB,
    PGPFB,
    EAGPF,
    QPR,
    WUP,
    EUP,
    QPRUP,
    HSEN,
    HSDIS,
}

There are 18 recognized commands for the flash. These are "bare-bones" commands and values that are written to the Flash's command register to inform the flash what to do. Table 14-5.

Variants
NOPWPEPCPBLPUPEAWGPBEGPBSSBPGPFBEAGPFQPRWUPEUPQPRUPHSENHSDIS