Enum sam4l::gpio::PeripheralFunction

pub enum PeripheralFunction {
    A,
    B,
    C,
    D,
    E,
    F,
    G,
}

Peripheral functions that may be assigned to a GPIOPin.

GPIO pins on the SAM4L may serve multiple functions. In addition to the default functionality, each pin can be assigned up to eight different peripheral functions. The various functions for each pin are described in "Peripheral Multiplexing I/O Lines" section of the SAM4L datasheet1.

Section 3.2, pages 19-29 ↩

Variants
ABCDEFG