Struct sam4l::gpio::GPIOPin

pub struct GPIOPin {
    port: *mut Registers,
    nvic: NvicIdx,
    pin_mask: u32,
    client_data: Cell<usize>,
    client: Cell<Option<&'static Client>>,
}

Fields
port: *mut Registers nvic: NvicIdx pin_mask: u32 client_data: Cell<usize> client: Cell<Option<&'static Client>>
Methods

impl GPIOPin

const fn new(pin: Pin, nvic: NvicIdx) -> GPIOPin

fn set_client<C: Client>(&self, client: &'static C)

fn select_peripheral(&self, function: PeripheralFunction)

fn enable(&self)

fn disable(&self)

fn enable_output(&self)

fn disable_output(&self)

fn enable_pull_down(&self)

fn disable_pull_down(&self)

fn enable_pull_up(&self)

fn disable_pull_up(&self)

fn set_interrupt_mode(&self, mode: u8)

Sets the interrupt mode registers. Interrupts may fire on the rising or falling edge of the pin or on both.

The mode is a two-bit value based on the mapping from section 23.7.13 of the SAM4L datasheet (page 563):

mode value 	Interrupt Mode

0b00 	Pin change

0b01 	Rising edge

0b10 	Falling edge

fn enable_interrupt(&self)

fn disable_interrupt(&self)

fn handle_interrupt(&self)

fn disable_schmidtt_trigger(&self)

fn enable_schmidtt_trigger(&self)

fn read(&self) -> bool

fn toggle(&self)

fn set(&self)

fn clear(&self)