Struct sam4l::pm::PowerManager

pub struct PowerManager {
    system_frequency: Cell<u32>,
    system_clock_source: Cell<SystemClockSource>,
}

Contains state for the power management peripheral. This includes the configurations for various system clocks and the final frequency that the system is running at.

Fields
system_frequency: Cell<u32>

Frequency at which the system clock is running.

system_clock_source: Cell<SystemClockSource>

Clock source configuration

Methods

impl PowerManager

unsafe fn setup_system_clock(&self, clock_source: SystemClockSource)

Sets up the system clock. This should be called as one of the first lines in the reset_handler within the platform's main.rs.