Struct sam4l::spi::Spi

pub struct Spi {
    registers: *mut SpiRegisters,
    client: Cell<Option<&'static SpiMasterClient>>,
    dma_read: Cell<Option<&'static DMAChannel>>,
    dma_write: Cell<Option<&'static DMAChannel>>,
    transfers_in_progress: Cell<u8>,
    dma_length: Cell<usize>,
    slave_client: Cell<Option<&'static SpiSlaveClient>>,
    role: Cell<SpiRole>,
}

The SAM4L supports four peripherals.

Fields
registers: *mut SpiRegisters client: Cell<Option<&'static SpiMasterClient>> dma_read: Cell<Option<&'static DMAChannel>> dma_write: Cell<Option<&'static DMAChannel>> transfers_in_progress: Cell<u8> dma_length: Cell<usize> slave_client: Cell<Option<&'static SpiSlaveClient>> role: Cell<SpiRole>
Methods

impl Spi

const fn new() -> Spi

Creates a new SPI object, with peripheral 0 selected

fn init_as_role(&self, role: SpiRole)

fn enable(&self)

fn disable(&self)

fn set_baud_rate(&self, rate: u32) -> u32

Sets the approximate baud rate for the active peripheral, and return the actual baud rate set.

Since the only supported baud rates are (system clock / n) where n is an integer from 1 to 255, the exact baud rate may not be available. In that case, the next lower baud rate will be selected.

The lowest available baud rate is 188235 baud. If the requested rate is lower, 188235 baud will be selected.

fn get_baud_rate(&self) -> u32

fn set_clock(&self, polarity: ClockPolarity)

fn get_clock(&self) -> ClockPolarity

fn set_phase(&self, phase: ClockPhase)

fn get_phase(&self) -> ClockPhase

fn set_active_peripheral(&self, peripheral: Peripheral)

fn get_active_peripheral(&self) -> Peripheral

Returns the currently active peripheral

fn read_active_csr(&self) -> u32

Returns the value of CSR0, CSR1, CSR2, or CSR3, whichever corresponds to the active peripheral

fn write_active_csr(&self, value: u32)

Sets the Chip Select Register (CSR) of the active peripheral (CSR0, CSR1, CSR2, or CSR3).

fn set_dma(&mut self, read: &'static DMAChannel, write: &'static DMAChannel)

Set the DMA channels used for reading and writing.

fn enable_clock(&self)

fn handle_interrupt(&self)

fn read_write_bytes(
&self,
write_buffer: Option<&'static mut [u8]>,
read_buffer: Option<&'static mut [u8]>,
len: usize
) -> ReturnCode

Asynchronous buffer read/write of SPI. returns SUCCESS if operation starts (will receive callback through SpiMasterClient), returns EBUSY if the operation does not start.